Restructures acc. to comments on PR
